The class main.dart was splited in new files
