+++ lib/input_with_icon.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class InputWithIcon extends StatefulWidget {
+  final IconData icon;
+  final String hint;
+  InputWithIcon({this.icon, this.hint});
+  @override
+  _InputWithIconState createState() => _InputWithIconState();
+}
+
+class _InputWithIconState extends State<InputWithIcon> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+          border: Border.all(color: Color(0xFFBC7C7C7), width: 1),
+          borderRadius: BorderRadius.circular(50)),
+      child: Row(
+        children: <Widget>[
+          Container(
+              width: 60,
+              child: Icon(
+                widget.icon,
+                size: 20,
+                color: Color(0xFFBB9B9B9),
+              )),
+          Expanded(
+            child: TextField(
+              decoration: InputDecoration(
+                  hintText: widget.hint,
+                  contentPadding: EdgeInsets.symmetric(vertical: 20)),
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
